Simplify MediaController access for No-Composable classes
